requirements.txt
pandas
gspread
shillelagh[all]
google-api-python-client
google-auth-oauthlib
streamlit-aggrid